update grpc client to use new keystore service api
